draft of moving ApiWorkflowClient.export to a separate file
