+++ lightly/api/api_workflow_client.py
@@ -14,6 +14,7 @@ from lightly.api.api_workflow_compute_worker import _ComputeWorkerMixin
 from lightly.api.api_workflow_datasets import _DatasetsMixin
 from lightly.api.api_workflow_datasources import _DatasourcesMixin
 from lightly.api.api_workflow_download_dataset import _DownloadDatasetMixin
+from lightly.api.api_workflow_download_export import _ExportDatasetMixin
 from lightly.api.api_workflow_predictions import _PredictionsMixin
 from lightly.api.api_workflow_selection import _SelectionMixin
 from lightly.api.api_workflow_tags import _TagsMixin
@@ -69,6 +70,7 @@ class ApiWorkflowClient(
     _CollaborationMixin,
     _PredictionsMixin,
     _ArtifactsMixin,
+    _ExportDatasetMixin,
 ):
     """Provides a uniform interface to communicate with the Lightly API.
